Testing

+++ controllers/message.js
@@ -42,8 +42,8 @@ async function addEmoji(message, icons, waitReact = false) {
 async function checkReaction(reaction, user) {
 	if (user.bot) return;
 	if (reaction.emoji.name === "❌" && 
-		reaction.message.deletable) {
-		console.log(reaction.message.author);
+		reaction.message.deletable &&
+		reaction.message.author.username === "PunishMeThanos") {
 		reaction.message.delete();
 	}
 	return;
